+++ lib/util/multi_select_item.dart
@@ -1,3 +1,4 @@
+/// A model class used to represent a selectable item.
 class MultiSelectItem<V> {
   const MultiSelectItem(this.value, this.label);